+++ tensorflow_datasets/core/logging/__init__.py
@@ -14,18 +14,21 @@
 # limitations under the License.
 
 """TFDS logging module."""
-from typing import Callable, List, Optional, TypeVar
+from typing import Callable, Dict, List, Optional, Tuple, TypeVar
 
 from absl import flags
 from tensorflow_datasets.core.logging import base_logger
+from tensorflow_datasets.core.logging import call_metadata
 from tensorflow_datasets.core.logging import logging_logger
 
 import wrapt
 
 
-_T = TypeVar('_T')
+_T = TypeVar("_T")
 
 _registered_loggers: Optional[List[base_logger.Logger]] = None
+
+_thread_ids_running_builder_init = set()
 
 
 def _check_init_registered_loggers() -> None:
@@ -52,32 +55,192 @@ def register(logger: base_logger.Logger) -> None:
   Args:
     logger: the logger to register.
   """
-  global _registered_loggers
   _check_init_registered_loggers()
   _registered_loggers.append(logger)
+
+
+def builder_init() -> Callable[[_T], _T]:
+  """"Decorator to call `builder_init` method on registered loggers."""
+
+  @wrapt.decorator
+  def decorator(function, dsbuilder, args, kwargs):
+    metadata = call_metadata.CallMetadata()
+    first_builder_init_in_stack = (
+        metadata.thread_id not in _thread_ids_running_builder_init)
+    if first_builder_init_in_stack:
+      _thread_ids_running_builder_init.add(metadata.thread_id)
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      if first_builder_init_in_stack:
+        metadata.mark_end()
+        _thread_ids_running_builder_init.remove(metadata.thread_id)
+        data_dir = kwargs.get("data_dir")
+        config = kwargs.get("config")
+        if config is not None:
+          config = str(config)
+        version = kwargs.get("version")
+        if version is not None:
+          version = str(version)
+        for logger in _get_registered_loggers():
+          logger.builder_init(
+              metadata=metadata,
+              name=dsbuilder.name,
+              data_dir=data_dir,
+              config=config,
+              version=version)
+
+  return decorator
+
+
+def _get_name_config_version_datadir(dsbuilder):
+  """Returns (builder_name, config, version, data_dir).
+
+  Args:
+    dsbuilder: the builder instance to get info for.
+  """
+  config_name = dsbuilder.builder_config.name if dsbuilder.builder_config else ""
+  data_path = dsbuilder.data_dir
+  return dsbuilder.name, config_name, str(dsbuilder.version), data_path
+
+
+def builder_info() -> Callable[[_T], _T]:
+  """"Decorator to call `builder_info` method on registered loggers."""
+
+  @wrapt.decorator
+  def decorator(function, dsbuilder, args, kwargs):
+    dsbuilder = args[0]  # Because property decorator applied first.
+    name, config_name, version, data_path = (
+        _get_name_config_version_datadir(dsbuilder))
+    metadata = call_metadata.CallMetadata()
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      metadata.mark_end()
+      for logger in _get_registered_loggers():
+        logger.builder_info(
+            metadata=metadata,
+            name=name,
+            config_name=config_name,
+            version=version,
+            data_path=data_path,
+        )
+
+  return decorator
 
 
 def as_dataset() -> Callable[[_T], _T]:
   """Decorator to call `as_dataset` method on registered loggers."""
 
   @wrapt.decorator
-  def decorator(function, builder, args, kwargs):
-    config_name = builder.builder_config.name if builder.builder_config else ''
-    data_path = builder.data_dir
+  def decorator(function, dsbuilder, args, kwargs):
+    name, config_name, version, data_path = (
+        _get_name_config_version_datadir(dsbuilder))
+    metadata = call_metadata.CallMetadata()
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      metadata.mark_end()
+      for logger in _get_registered_loggers():
+        logger.as_dataset(
+            metadata=metadata,
+            name=name,
+            config_name=config_name,
+            version=version,
+            data_path=data_path,
+            split=args and args[0] or kwargs.get("split"),
+            batch_size=kwargs.get("batch_size"),
+            shuffle_files=kwargs.get("shuffle_files"),
+            read_config=kwargs.get("read_config"),
+            as_supervised=kwargs.get("as_supervised"),
+            decoders=kwargs.get("decoders"),
+        )
 
-    for logger in _get_registered_loggers():
-      logger.as_dataset(
-          dataset_name=builder.name,
-          config_name=config_name,
-          version=str(builder.version),
-          data_path=data_path,
-          split=args and args[0] or kwargs.get('split', 'all'),
-          batch_size=kwargs.get('batch_size'),
-          shuffle_files=kwargs.get('shuffle_files', False),
-          read_config=kwargs.get('read_config', None),
-          as_supervised=kwargs.get('as_supervised', False),
-          decoders=kwargs.get('decoders', None))
+  return decorator
 
-    return function(*args, **kwargs)
+
+def list_builders() -> Callable[[_T], _T]:
+  """Decorator to call `list_builders` method on registered loggers."""
+
+  @wrapt.decorator
+  def decorator(function, unused_none_instance, args, kwargs):
+    metadata = call_metadata.CallMetadata()
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      metadata.mark_end()
+      for logger in _get_registered_loggers():
+        logger.list_builders(
+            metadata=metadata,
+            with_community_datasets=kwargs.get("with_community_datasets"))
+
+  return decorator
+
+
+def load() -> Callable[[_T], _T]:
+  """Decorator to call `load` method on registered loggers."""
+
+  @wrapt.decorator
+  def decorator(function, unused_none_instance, args, kwargs):
+    metadata = call_metadata.CallMetadata()
+    name = args[0] if args else kwargs["name"]
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      metadata.mark_end()
+      for logger in _get_registered_loggers():
+        logger.load(
+            metadata=metadata,
+            name=name,
+            split=kwargs.get("split"),
+            data_dir=kwargs.get("data_dir"),
+            batch_size=kwargs.get("batch_size"),
+            shuffle_files=kwargs.get("shuffle_files"),
+            download=kwargs.get("download"),
+            as_supervised=kwargs.get("as_supervised"),
+            decoders=kwargs.get("decoders"),
+            read_config=kwargs.get("read_config"),
+            with_info=kwargs.get("with_info"),
+            try_gcs=kwargs.get("try_gcs"),
+        )
+
+  return decorator
+
+
+def builder() -> Callable[[_T], _T]:
+  """Decorator to call `builder` method on registered loggers."""
+
+  @wrapt.decorator
+  def decorator(function, unused_none_instance, args, kwargs):
+    metadata = call_metadata.CallMetadata()
+    name = args[0] if args else kwargs["name"]
+    try:
+      return function(*args, **kwargs)
+    except Exception:
+      metadata.mark_error()
+      raise
+    finally:
+      metadata.mark_end()
+      for logger in _get_registered_loggers():
+        logger.builder(
+            metadata=metadata,
+            name=name,
+            try_gcs=kwargs.get("try_gcs"),
+        )
 
   return decorator